Prevented deletion with -debug

+++ regp.go
@@ -26,9 +26,11 @@ func main () {
 	lg.Log(lg.INFO, "Initiating parse job...")
 	for _, v := range flag.Args() {
 		parser(v)
-		err := os.Remove(v)
-		if err != nil {
-			lg.Log(lg.ERROR, "Failed to remove file")
+		if !*dbgmode {
+			err := os.Remove(v)
+			if err != nil {
+				lg.Log(lg.ERROR, "Failed to remove file")
+			}
 		}
 	}
 }
